README

+++ examples/MSMARCO/train_index_and_encoder.py
@@ -5,12 +5,9 @@ import faiss
 import pickle
 from transformers import HfArgumentParser
 
-from LibVQ.baseindex import FaissIndex
-from LibVQ.inference import inference
 from LibVQ.models import Encoder, EncoderConfig
 from LibVQ.dataset.dataset import load_rel, write_rel
 from LibVQ.learnable_index import LearnableIndex
-from LibVQ.dataset.preprocess import preprocess_data
 from LibVQ.utils import setuplogging
 
 from arguments import IndexArguments, DataArguments, ModelArguments, TrainingArguments
@@ -45,19 +42,21 @@ if __name__ == '__main__':
 
 
     # creat learnable_index
-    learnable_index = LearnableIndex(encoder=text_encoder,
+    learnable_index = LearnableIndex(index_method=index_args.index_method,
+                                     encoder=text_encoder,
                                      index_file=os.path.join(data_args.output_dir, f'{index_args.index_method}.index'))
 
-    # generate hard negative if need
-    '''
+    # The class randomly sample the negative from corpus by default. You also can assgin speficed negative for each query (set --neg_file)
     neg_file = os.path.join(data_args.output_dir, f"train-queries_hardneg.pickle")
-    train_ground_truths = load_rel(os.path.join(data_args.preprocess_dir, 'train-rels.tsv'))
-    trainquery2hardneg = learnable_index.hard_negative(train_query_embeddings, train_ground_truths, topk=400, batch_size=64)
-    pickle.dump(trainquery2hardneg, open(neg_file, 'wb'))
-    '''
+    if not os.path.exists(neg_file):
+        train_ground_truths = load_rel(os.path.join(data_args.preprocess_dir, 'train-rels.tsv'))
+        trainquery2hardneg = learnable_index.hard_negative(train_query_embeddings, train_ground_truths, topk=400,
+                                                           batch_size=64)
+        pickle.dump(trainquery2hardneg, open(neg_file, 'wb'))
+
 
     # contrastive learning
-    if training_args.training_mode == 'contrastive':
+    if training_args.training_mode == 'contrastive_jointly':
         data_args.save_ckpt_dir = f'./saved_ckpts/{training_args.training_mode}/'
         learnable_index.fit_with_multi_gpus(rel_file=os.path.join(data_args.preprocess_dir, 'train-rels.tsv'),
                                             neg_file=os.path.join(data_args.output_dir, f"train-queries_hardneg.pickle"),
@@ -76,7 +75,7 @@ if __name__ == '__main__':
 
 
     # distill learning
-    if training_args.training_mode == 'distill':
+    if training_args.training_mode == 'distill_jointly':
         data_args.save_ckpt_dir = f'./saved_ckpts/{training_args.training_mode}/'
         learnable_index.fit_with_multi_gpus(rel_file=os.path.join(data_args.preprocess_dir, 'train-rels.tsv'),
                                             neg_file=os.path.join(data_args.output_dir, f"train-queries_hardneg.pickle"),
@@ -94,16 +93,45 @@ if __name__ == '__main__':
                                             fix_emb='doc',
                                             epochs=30)
 
-    # distill with no label data
-    if training_args.training_mode == 'distill_nolabel':
-        '''
-         # generate train data from retrieve results if need
-         # query2pos, query2neg = trainquery2hardneg = learnable_index.generate_virtual_traindata(train_query_embeddings, topk=400, batch_size=64)
-         # write_rel(os.path.join(data_args.output_dir, 'train-virtual_rel.tsv'), query2pos)
-         # pickle.dump(query2neg, open(os.path.join(data_args.output_dir, f"train-queries-virtual_hardneg.pickle"), 'wb'))
-         '''
 
-        data_args.save_ckpt_dir = f'./saved_ckpts/distill_nolabel_{training_args.per_device_train_batch_size}/'
+    # distill learning and train both query encoder and doc encoder, which only can be used when ivf is disabled
+    if training_args.training_mode == 'distill_jointly_v2':
+        data_args.save_ckpt_dir = f'./saved_ckpts/{training_args.training_mode}/'
+        assert  'ivf' not in index_args.index_method
+        learnable_index.fit_with_multi_gpus(rel_file=os.path.join(data_args.preprocess_dir, 'train-rels.tsv'),
+                                            neg_file=os.path.join(data_args.output_dir, f"train-queries_hardneg.pickle"),
+                                            query_data_dir=data_args.preprocess_dir,
+                                            max_query_length=data_args.max_query_length,
+                                            doc_data_dir=data_args.preprocess_dir,
+                                            max_doc_length=128,
+                                            query_embeddings_file=data_args.query_embeddings_file,
+                                            doc_embeddings_file=data_args.doc_embeddings_file,
+                                            emb_size = emb_size,
+                                            checkpoint_path=data_args.save_ckpt_dir,
+                                            logging_steps=training_args.logging_steps,
+                                            per_device_train_batch_size=training_args.per_device_train_batch_size,
+                                            loss_weight={'encoder_weight': 1.0, 'pq_weight': 1.0, 'ivf_weight': 0.0},
+                                            lr_params={'encoder_lr': 1e-5, 'pq_lr': 1e-4, 'ivf_lr': 0.0},
+                                            loss_method='distill',
+                                            fix_emb=None,
+                                            epochs=30)
+
+    # distill with no label data
+    if training_args.training_mode == 'distill_virtual-data_jointly':
+
+        # generate train data by brute-force or the index which should has similar performance with brute force
+        if not os.path.exists(os.path.join(data_args.output_dir, 'train-virtual_rel.tsv')):
+            # flat_index = FaissIndex(doc_embeddings=doc_embeddings, index_method='flat', dist_mode='ip')
+            # query2pos, query2neg = trainquery2hardneg = flat_index.generate_virtual_traindata(train_query_embeddings,
+            #                                                                                        topk=400, batch_size=64)
+            # or
+            query2pos, query2neg = trainquery2hardneg = learnable_index.generate_virtual_traindata(train_query_embeddings, topk=400, batch_size=64)
+
+            write_rel(os.path.join(data_args.output_dir, 'train-virtual_rel.tsv'), query2pos)
+            pickle.dump(query2neg, open(os.path.join(data_args.output_dir, f"train-queries-virtual_hardneg.pickle"), 'wb'))
+
+
+        data_args.save_ckpt_dir = f'./saved_ckpts/{training_args.training_mode}_{training_args.per_device_train_batch_size}/'
         learnable_index.fit_with_multi_gpus(rel_file=os.path.join(data_args.output_dir, 'train-virtual_rel.tsv'),
                                             neg_file=os.path.join(data_args.output_dir, f"train-queries-virtual_hardneg.pickle"),
                                             query_data_dir=data_args.preprocess_dir,
@@ -111,7 +139,7 @@ if __name__ == '__main__':
                                             query_embeddings_file=data_args.query_embeddings_file,
                                             doc_embeddings_file=data_args.doc_embeddings_file,
                                             emb_size = emb_size,
-                                            per_query_neg_num = 200,
+                                            per_query_neg_num = 1,
                                             checkpoint_path=data_args.save_ckpt_dir,
                                             logging_steps=training_args.logging_steps,
                                             per_device_train_batch_size=training_args.per_device_train_batch_size,
@@ -119,14 +147,10 @@ if __name__ == '__main__':
                                             lr_params={'encoder_lr': 1e-5, 'pq_lr': 1e-4, 'ivf_lr': 1e-3},
                                             loss_method='distill',
                                             fix_emb='doc',
-                                            epochs=2)
+                                            epochs=30)
 
-    # select a ckpt
-    # or set saved_ckpts_path in function "update_encoder“ and "update_index_with_ckpt",  which will select the latest ckpt to update
+    # select a latest ckpt
     ckpt_path = learnable_index.get_latest_ckpt(data_args.save_ckpt_dir)
-    print(ckpt_path)
-
-    # ckpt_path = './saved_ckpts/distill_512/epoch_28_step_3567'
 
     # update query embeddings when re-training the query encoder
     data_args.output_dir = f'./data/passage/evaluate/LearnableIndex_{training_args.loss_method}_{training_args.fix_emb}'
@@ -143,9 +167,6 @@ if __name__ == '__main__':
                                  mode="r")
     new_query_embeddings = new_query_embeddings.reshape(-1, emb_size)
 
-    diff = np.sum((new_query_embeddings - query_embeddings)**2)
-    print(diff)
-
 
     # update index
     print('Updating the index with new ivf and pq')
@@ -153,17 +174,6 @@ if __name__ == '__main__':
 
     # Test
     ground_truths = load_rel(os.path.join(data_args.preprocess_dir, 'dev-rels.tsv'))
-    # learnable_index.set_nprobe(index_args.nprobe)
-    # qids = list(range(len(new_query_embeddings)))
-    # learnable_index.test(new_query_embeddings, qids, ground_truths, topk=1000, batch_size=64,
-    #                      MRR_cutoffs=[10, 100], Recall_cutoffs=[10, 30, 50, 100])
-
-    learnable_index.set_nprobe(1)
-    qids = list(range(len(new_query_embeddings)))
-    learnable_index.test(new_query_embeddings, qids, ground_truths, topk=1000, batch_size=64,
-                         MRR_cutoffs=[10, 100], Recall_cutoffs=[10, 30, 50, 100])
-
-    learnable_index.set_nprobe(100)
-    qids = list(range(len(new_query_embeddings)))
-    learnable_index.test(new_query_embeddings, qids, ground_truths, topk=1000, batch_size=64,
-                         MRR_cutoffs=[10, 100], Recall_cutoffs=[5, 10, 20, 30, 50, 100])
+    learnable_index.test(new_query_embeddings, ground_truths, topk=1000, batch_size=64,
+                         MRR_cutoffs=[10, 100], Recall_cutoffs=[10, 30, 50, 100],
+                         nprobe=index_args.nprobe)
